Split enclave functions from pal into separate header
