+++ src/enclave/main.cpp
@@ -3,7 +3,9 @@
 #include "ccf/ds/ccf_exception.h"
 #include "ccf/ds/json.h"
 #include "ccf/ds/logger.h"
-#include "ccf/pal.h"
+#include "ccf/pal/enclave.h"
+#include "ccf/pal/locking.h"
+#include "ccf/pal/mem.h"
 #include "ccf/version.h"
 #include "common/enclave_interface_types.h"
 #include "enclave.h"
